requirements.txt
aiohttp<=3.8.3
anytree<=2.8.0
aria2p<=0.11.3
appdirs<=1.4.4
attrdict<=2.0.1
beautifulsoup4<=4.10.0
cloudscraper<=1.2.66
feedparser<=6.0.10
git+https://github.com/breakdowns/lk21
google-api-python-client<=2.70.0
google-auth-httplib2
google-auth-oauthlib
gunicorn<=20.1.0
hachoir
js2py
lxml<=4.8.0
pillow<=9.3.0
psutil
psycopg2-binary
pybase64
pyrogram==1.4.16
pyshorteners
python-dotenv
python-magic
python-telegram-bot==13.15
qbittorrent-api
requests
speedtest-cli
telegraph<=2.2.0
tenacity
TgCrypto
torrentool==1.1.0
urllib3<=1.26.15
youtube_dl
